template/requirements.txt
altgraph==0.17
aniso8601==7.0.0
appdirs==1.4.4
APScheduler==3.7.0
attrs==20.3.0
bcrypt==3.2.0
bidict==0.21.2
blinker==1.4
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==7.1.2
Deprecated==1.2.10
dnspython==1.16.0
email-validator==1.1.2
exponent-server-sdk==2.0.0
filelock==3.0.12
Flask==1.1.2
Flask-APScheduler==1.12.2
Flask-Bcrypt==0.7.1
Flask-Cors==3.0.10
Flask-DebugToolbar==0.11.0
Flask-GraphQL==2.0.1
# Flask-GraphQL-Auth==1.3.3
Flask-JWT-Extended==3.25.0
Flask-Limiter==1.4
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-Mobility==1.1.0
Flask-RESTful==0.3.8
flask-restful-swagger-3==0.4.3
flask-restx==0.2.0
Flask-SocketIO==5.0.1
Flask-SQLAlchemy==2.4.4
Flask-WTF==0.14.3
flaskwebgui==0.2.3
future==0.18.2
gevent==21.1.2
graphene==2.1.9
graphene-sqlalchemy==2.3.0
graphql-core==2.3.2
graphql-relay==2.0.1
graphql-server-core==1.2.0
greenlet==1.0.0
idna==2.10
importlib-metadata==3.9.0
itsdangerous==1.1.0
Jinja2==2.11.2
jsonschema==3.2.0
limits==1.5.1
macholib==1.15
MarkupSafe==1.1.1
numpy==1.20.3
pandas==1.2.4
pefile==2019.4.18
promise==2.3
psutil==5.8.0
pycparser==2.20
pyinstaller==4.2
pyinstaller-hooks-contrib==2021.1
PyJWT==1.7.1
pyodbc==4.0.30
pyrsistent==0.17.3
python-dateutil==2.8.1
python-engineio==4.0.1
python-socketio==5.1.0
pytz==2020.5
pywin32-ctypes==0.2.0
requests==2.25.1
Rx==1.6.1
singledispatch==3.7.0
six==1.15.0
SQLAlchemy==1.3.22
SQLAlchemy-Utils==0.36.8
tinyaes==1.0.1
typing-extensions==3.7.4.3
tzlocal==2.1
urllib3==1.26.2
waitress==2.0.0
Werkzeug==1.0.1
whichcraft==0.6.1
wrapt==1.12.1
WTForms==2.3.3
zipp==3.4.1
zope.event==4.5.0
zope.interface==5.3.0
